Remove usage of the TOPOLOGY environment variable in tests.
GODRIVER-425
