+++ examples/documentation_examples/examples_test.go
@@ -11,7 +11,6 @@ package documentation_examples_test
 
 import (
 	"context"
-	"os"
 	"testing"
 	"time"
 
@@ -21,6 +20,9 @@ import (
 	"go.mongodb.org/mongo-driver/mongo"
 	"go.mongodb.org/mongo-driver/mongo/options"
 	"go.mongodb.org/mongo-driver/x/bsonx"
+	"go.mongodb.org/mongo-driver/x/mongo/driver/connstring"
+	"go.mongodb.org/mongo-driver/x/mongo/driver/description"
+	"go.mongodb.org/mongo-driver/x/mongo/driver/topology"
 )
 
 func TestDocumentationExamples(t *testing.T) {
@@ -48,13 +50,13 @@ func TestDocumentationExamples(t *testing.T) {
 func TestTransactionExamples(t *testing.T) {
 	ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
 	defer cancel()
-	cs := testutil.ConnString(t)
-	client, err := mongo.Connect(context.Background(), options.Client().ApplyURI(cs.String()))
+	topo := createTopology(t)
+	client, err := mongo.Connect(context.Background(), &options.ClientOptions{Deployment: topo})
 	require.NoError(t, err)
 	defer client.Disconnect(ctx)
 
 	ver, err := getServerVersion(ctx, client)
-	if err != nil || testutil.CompareVersions(t, ver, "4.0") < 0 || os.Getenv("TOPOLOGY") != "replica_set" {
+	if err != nil || testutil.CompareVersions(t, ver, "4.0") < 0 || topo.Kind() != description.ReplicaSet {
 		t.Skip("server does not support transactions")
 	}
 	err = documentation_examples.TransactionsExamples(ctx, client)
@@ -64,14 +66,14 @@ func TestTransactionExamples(t *testing.T) {
 func TestChangeStreamExamples(t *testing.T) {
 	ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
 	defer cancel()
-	cs := testutil.ConnString(t)
-	client, err := mongo.Connect(context.Background(), options.Client().ApplyURI(cs.String()))
+	topo := createTopology(t)
+	client, err := mongo.Connect(context.Background(), &options.ClientOptions{Deployment: topo})
 	require.NoError(t, err)
 	defer client.Disconnect(ctx)
 
 	db := client.Database("changestream_examples")
 	ver, err := getServerVersion(ctx, client)
-	if err != nil || testutil.CompareVersions(t, ver, "3.6") < 0 || os.Getenv("TOPOLOGY") != "replica_set" {
+	if err != nil || testutil.CompareVersions(t, ver, "3.6") < 0 || topo.Kind() != description.ReplicaSet {
 		t.Skip("server does not support changestreams")
 	}
 	documentation_examples.ChangeStreamExamples(t, db)
@@ -92,4 +94,14 @@ func getServerVersion(ctx context.Context, client *mongo.Client) (string, error)
 	}
 
 	return version.StringValue(), nil
+}
+
+func createTopology(t *testing.T) *topology.Topology {
+	topo, err := topology.New(topology.WithConnString(func(connstring.ConnString) connstring.ConnString {
+		return testutil.ConnString(t)
+	}))
+	if err != nil {
+		t.Fatalf("topology.New error: %v", err)
+	}
+	return topo
 }
